Fixed claim success dialog getting hidden
