+++ apps/scoutgame/lib/points/claimPointsAction.ts
@@ -8,5 +8,6 @@ import { authActionClient } from 'lib/actions/actionClient';
 export const claimPointsAction = authActionClient.metadata({ actionName: 'claim_points' }).action(async ({ ctx }) => {
   const result = await claimPoints({ userId: ctx.session.scoutId });
   revalidatePath('/profile');
+  revalidatePath('/claim');
   return { success: true, claimedPoints: result.total };
 });
